+++ NeoHubRLib/Shared.swift
@@ -1,11 +1,5 @@
 import Foundation
 
-public struct Socket {
-    public static var addr: String {
-        "/tmp/neohubr-\(getuid()).sock"
-    }
-}
-
 public struct RunRequest: Codable, Sendable {
     public let wd: URL
     public let bin: URL
@@ -28,43 +22,6 @@ public struct RunRequest: Codable, Sendable {
         self.path = path
         self.opts = opts
         self.env = env
-    }
-}
-
-public struct CLIErrorReport: Codable, Sendable {
-    public let message: String
-    public let detail: String?
-    public let code: Int?
-
-    public init(message: String, detail: String? = nil, code: Int? = nil) {
-        self.message = message
-        self.detail = detail
-        self.code = code
-    }
-}
-
-public enum IPCMessageType: String, Codable, Sendable {
-    case run
-    case cliError
-}
-
-public struct IPCMessage: Codable, Sendable {
-    public let type: IPCMessageType
-    public let run: RunRequest?
-    public let cliError: CLIErrorReport?
-
-    private init(type: IPCMessageType, run: RunRequest?, cliError: CLIErrorReport?) {
-        self.type = type
-        self.run = run
-        self.cliError = cliError
-    }
-
-    public static func run(_ request: RunRequest) -> IPCMessage {
-        IPCMessage(type: .run, run: request, cliError: nil)
-    }
-
-    public static func cliError(_ report: CLIErrorReport) -> IPCMessage {
-        IPCMessage(type: .cliError, run: nil, cliError: report)
     }
 }